Move parser.py to parser/ in preparation for new parsers
